add basePath for dashboard
