+++ packages/dashboard/src/pages/journeys/[id]/index.page.tsx
@@ -1,0 +1,23 @@
+import { useRouter } from "next/router";
+
+import JourneysBuilder from "../../../components/journeys/journeysBuilder";
+import JourneyLayout from "../../../components/journeys/layout";
+import {
+  JourneyGetServerSideProps,
+  journeyGetServerSideProps,
+} from "./getServerSideProps";
+
+export const getServerSideProps: JourneyGetServerSideProps = (ctx) =>
+  journeyGetServerSideProps(ctx);
+
+function Journey() {
+  const path = useRouter();
+  const id = typeof path.query.id === "string" ? path.query.id : undefined;
+
+  return (
+    <JourneyLayout journeyId={id}>
+      <JourneysBuilder />
+    </JourneyLayout>
+  );
+}
+export default Journey;
